move input_line from command_line to simplewallet
It was only used there, and this removes one part of the common
dependency on libreadline
