+++ src/common/command_line.h
@@ -41,8 +41,6 @@
 namespace command_line
 {
 
-  std::string input_line(const std::string& prompt);
-
   //! \return True if `str` is `is_iequal("y" || "yes" || `tr("yes"))`.
   bool is_yes(const std::string& str);
   //! \return True if `str` is `is_iequal("n" || "no" || `tr("no"))`.
